requirements.txt

azure-ai-evaluation[redteam]>=1.16.4
azure-identity
openai>=1.108.1
python-dotenv
rich
prompty
